requirement.txt
flask
pymongo
python-dotenv
dnspython
certifi